requirements.txt
streamlit
pandas
altair
fpdf
google-genai
Pillow